Move review prompt to blocks
